requirements.txt
pandas==1.3.5
streamlit==1.8.1
matplotlib==3.5.1
seaborn==0.11.2
numpy
Pillow==9.0.0
requests==2.26.0
web3==5.29.2
python-dotenv==0.19.1
shroomdk
